requirements.txt
ultralytics
streamlit
opencv-python
numpy
pillow
streamlit-webrtc
torch
torchvision
pyyaml